console: Refactor add device form into container
