+++ pkg/webui/console/containers/device-data-form/messages.js
