+++ pkg/webui/console/containers/device-data-form/validation-schema.js
