+++ pkg/webui/locales/en.json
@@ -24,6 +24,32 @@
   "console.containers.applications-table.index.desc": "Description",
   "console.containers.applications-table.index.empty": "No items matched your criteria",
   "console.containers.collaborators-table.index.id": "User / Organization ID",
+  "console.containers.device-data-form.messages.activationSettings": "Activation Settings",
+  "console.containers.device-data-form.messages.createDevice": "Create Device",
+  "console.containers.device-data-form.messages.deviceIdPlaceholder": "my-new-device",
+  "console.containers.device-data-form.messages.deviceNamePlaceholder": "My New Device",
+  "console.containers.device-data-form.messages.deviceIdDescription": "Unique Identifier of your device; this cannot be changed afterwards",
+  "console.containers.device-data-form.messages.deviceNameDescription": "Human friendly name of your device for display purposes",
+  "console.containers.device-data-form.messages.deviceDescDescription": "Optional device description; can also be used to save notes about the device",
+  "console.containers.device-data-form.messages.joinEUIPlaceholder": "The connected Join EUI",
+  "console.containers.device-data-form.messages.leaveBlankPlaceholder": "Leave blank to generate automatically",
+  "console.containers.device-data-form.messages.resetsJoinNonces": "Resets Join Nonces",
+  "console.containers.device-data-form.messages.resetsFCnt": "Resets Frame Counters",
+  "console.containers.device-data-form.messages.deviceEUIDescription": "The device EUI is the unique identifier for this device on the network. Can be changed later.",
+  "console.containers.device-data-form.messages.nwkKeyDescription": "The encrypted Network Key",
+  "console.containers.device-data-form.messages.appKeyDescription": "The App Key is used to secure the communication between your device and the network.",
+  "console.containers.device-data-form.messages.appSKeyDescription": "App Session Key",
+  "console.containers.device-data-form.messages.fwdNtwkKeyDescription": "Forwarding Network Session Integrity Key (or LoRaWAN 1.0.x NwkSKey)",
+  "console.containers.device-data-form.messages.sNtwkSIKeyDescription": "Serving Network Session Integrity Key (only for LoRaWAN 1.1+)",
+  "console.containers.device-data-form.messages.ntwkSEncKeyDescription": "Network Session Encryption Key (only for LoRaWAN 1.1+)",
+  "console.containers.device-data-form.messages.validate8": "This value needs to be exactly 8 characters long",
+  "console.containers.device-data-form.messages.validate16": "This value needs to be exactly 16 characters long",
+  "console.containers.device-data-form.messages.validate32": "This value needs to be exactly 32 characters long",
+  "console.containers.device-data-form.messages.supportsClassC": "Supports Class C",
+  "console.containers.device-data-form.messages.activationMode": "Activation Mode",
+  "console.containers.device-data-form.messages.otaa": "Over The Air Activation (OTAA)",
+  "console.containers.device-data-form.messages.abp": "Activation By Personalization (ABP)",
+  "console.containers.device-data-form.messages.resetWarning": "Reseting is insecure and makes your device susceptible for replay attacks",
   "console.containers.devices-table.index.deviceId": "Device ID",
   "console.containers.devices-table.index.connectedDevices": "Connected Devices ({deviceCount})",
   "console.containers.devices-table.index.add": "Add Device",
@@ -64,32 +90,6 @@
   "console.views.application-link.index.statistics": "Statistics",
   "console.views.application-link.index.unlink": "Unlink",
   "console.views.application-link.index.unlinkSuccess": "Successfully unlinked",
-  "console.views.device-add.messages.activationSettings": "Activation Settings",
-  "console.views.device-add.messages.createDevice": "Create Device",
-  "console.views.device-add.messages.deviceIdPlaceholder": "my-new-device",
-  "console.views.device-add.messages.deviceNamePlaceholder": "My New Device",
-  "console.views.device-add.messages.deviceIdDescription": "Unique Identifier of your device; this cannot be changed afterwards",
-  "console.views.device-add.messages.deviceNameDescription": "Human friendly name of your device for display purposes",
-  "console.views.device-add.messages.deviceDescDescription": "Optional device description; can also be used to save notes about the device",
-  "console.views.device-add.messages.joinEUIPlaceholder": "The connected Join EUI",
-  "console.views.device-add.messages.leaveBlankPlaceholder": "Leave blank to generate automatically",
-  "console.views.device-add.messages.resetsJoinNonces": "Resets Join Nonces",
-  "console.views.device-add.messages.resetsFCnt": "Resets Frame Counters",
-  "console.views.device-add.messages.deviceEUIDescription": "The device EUI is the unique identifier for this device on the network. Can be changed later.",
-  "console.views.device-add.messages.nwkKeyDescription": "The encrypted Network Key",
-  "console.views.device-add.messages.appKeyDescription": "The App Key is used to secure the communication between your device and the network.",
-  "console.views.device-add.messages.appSKeyDescription": "App Session Key",
-  "console.views.device-add.messages.fwdNtwkKeyDescription": "Forwarding Network Session Integrity Key (or LoRaWAN 1.0.x NwkSKey)",
-  "console.views.device-add.messages.sNtwkSIKeyDescription": "Serving Network Session Integrity Key (only for LoRaWAN 1.1+)",
-  "console.views.device-add.messages.ntwkSEncKeyDescription": "Network Session Encryption Key (only for LoRaWAN 1.1+)",
-  "console.views.device-add.messages.validate8": "This value needs to be exactly 8 characters long",
-  "console.views.device-add.messages.validate16": "This value needs to be exactly 16 characters long",
-  "console.views.device-add.messages.validate32": "This value needs to be exactly 32 characters long",
-  "console.views.device-add.messages.supportsClassC": "Supports Class C",
-  "console.views.device-add.messages.activationMode": "Activation Mode",
-  "console.views.device-add.messages.otaa": "Over The Air Activation (OTAA)",
-  "console.views.device-add.messages.abp": "Activation By Personalization (ABP)",
-  "console.views.device-add.messages.resetWarning": "Reseting is insecure and makes your device susceptible for replay attacks",
   "console.views.device-overview.index.hardware": "Hardware",
   "console.views.device-overview.index.brand": "Brand",
   "console.views.device-overview.index.model": "Model",
