+++ pkg/webui/locales/xx.json
@@ -24,6 +24,32 @@
   "console.containers.applications-table.index.desc": "Xxxxxxxxxxx",
   "console.containers.applications-table.index.empty": "Xx xxxxx xxxxxxx xxxx xxxxxxxx",
   "console.containers.collaborators-table.index.id": "Xxxx / Xxxxxxxxxxxx XX",
+  "console.containers.device-data-form.messages.activationSettings": "Xxxxxxxxxx Xxxxxxxx",
+  "console.containers.device-data-form.messages.createDevice": "Xxxxxx Xxxxxx",
+  "console.containers.device-data-form.messages.deviceIdPlaceholder": "xx-xxx-xxxxxx",
+  "console.containers.device-data-form.messages.deviceNamePlaceholder": "Xx Xxx Xxxxxx",
+  "console.containers.device-data-form.messages.deviceIdDescription": "Xxxxxx Xxxxxxxxxx xx xxxx xxxxxx; xxxx xxxxxx xx xxxxxxx xxxxxxxxxx",
+  "console.containers.device-data-form.messages.deviceNameDescription": "Xxxxx xxxxxxxx xxxx xx xxxx xxxxxx xxx xxxxxxx xxxxxxxx",
+  "console.containers.device-data-form.messages.deviceDescDescription": "Xxxxxxxx xxxxxx xxxxxxxxxxx; xxx xxxx xx xxxx xx xxxx xxxxx xxxxx xxx xxxxxx",
+  "console.containers.device-data-form.messages.joinEUIPlaceholder": "Xxx xxxxxxxxx Xxxx XXX",
+  "console.containers.device-data-form.messages.leaveBlankPlaceholder": "Xxxxx xxxxx xx xxxxxxxx xxxxxxxxxxxxx",
+  "console.containers.device-data-form.messages.resetsJoinNonces": "Xxxxxx Xxxx Xxxxxx",
+  "console.containers.device-data-form.messages.resetsFCnt": "Xxxxxx Xxxxx Xxxxxxxx",
+  "console.containers.device-data-form.messages.deviceEUIDescription": "Xxx xxxxxx XXX xx xxx xxxxxx xxxxxxxxxx xxx xxxx xxxxxx xx xxx xxxxxxx. Xxx xx xxxxxxx xxxxx.",
+  "console.containers.device-data-form.messages.nwkKeyDescription": "Xxx xxxxxxxxx Xxxxxxx Xxx",
+  "console.containers.device-data-form.messages.appKeyDescription": "Xxx Xxx Xxx xx xxxx xx xxxxxx xxx xxxxxxxxxxxxx xxxxxxx xxxx xxxxxx xxx xxx xxxxxxx.",
+  "console.containers.device-data-form.messages.appSKeyDescription": "Xxx Xxxxxxx Xxx",
+  "console.containers.device-data-form.messages.fwdNtwkKeyDescription": "Xxxxxxxxxx Xxxxxxx Xxxxxxx Xxxxxxxxx Xxx (xx XxXxXXX x.x.x XxxXXxx)",
+  "console.containers.device-data-form.messages.sNtwkSIKeyDescription": "Xxxxxxx Xxxxxxx Xxxxxxx Xxxxxxxxx Xxx (xxxx xxx XxXxXXX x.x+)",
+  "console.containers.device-data-form.messages.ntwkSEncKeyDescription": "Xxxxxxx Xxxxxxx Xxxxxxxxxx Xxx (xxxx xxx XxXxXXX x.x+)",
+  "console.containers.device-data-form.messages.validate8": "Xxxx xxxxx xxxxx xx xx xxxxxxx x xxxxxxxxxx xxxx",
+  "console.containers.device-data-form.messages.validate16": "Xxxx xxxxx xxxxx xx xx xxxxxxx xx xxxxxxxxxx xxxx",
+  "console.containers.device-data-form.messages.validate32": "Xxxx xxxxx xxxxx xx xx xxxxxxx xx xxxxxxxxxx xxxx",
+  "console.containers.device-data-form.messages.supportsClassC": "Xxxxxxxx Xxxxx X",
+  "console.containers.device-data-form.messages.activationMode": "Xxxxxxxxxx Xxxx",
+  "console.containers.device-data-form.messages.otaa": "Xxxx Xxx Xxx Xxxxxxxxxx (XXXX)",
+  "console.containers.device-data-form.messages.abp": "Xxxxxxxxxx Xx Xxxxxxxxxxxxxxx (XXX)",
+  "console.containers.device-data-form.messages.resetWarning": "Xxxxxxxx xx xxxxxxxx xxx xxxxx xxxx xxxxxx xxxxxxxxxxx xxx xxxxxx xxxxxxx",
   "console.containers.devices-table.index.deviceId": "Xxxxxx XX",
   "console.containers.devices-table.index.connectedDevices": "Xxxxxxxxx Xxxxxxx ({deviceCount})",
   "console.containers.devices-table.index.add": "Xxx Xxxxxx",
@@ -64,32 +90,6 @@
   "console.views.application-link.index.statistics": "Xxxxxxxxxx",
   "console.views.application-link.index.unlink": "Xxxxxx",
   "console.views.application-link.index.unlinkSuccess": "Xxxxxxxxxxxx xxxxxxxx",
-  "console.views.device-add.messages.activationSettings": "Xxxxxxxxxx Xxxxxxxx",
-  "console.views.device-add.messages.createDevice": "Xxxxxx Xxxxxx",
-  "console.views.device-add.messages.deviceIdPlaceholder": "xx-xxx-xxxxxx",
-  "console.views.device-add.messages.deviceNamePlaceholder": "Xx Xxx Xxxxxx",
-  "console.views.device-add.messages.deviceIdDescription": "Xxxxxx Xxxxxxxxxx xx xxxx xxxxxx; xxxx xxxxxx xx xxxxxxx xxxxxxxxxx",
-  "console.views.device-add.messages.deviceNameDescription": "Xxxxx xxxxxxxx xxxx xx xxxx xxxxxx xxx xxxxxxx xxxxxxxx",
-  "console.views.device-add.messages.deviceDescDescription": "Xxxxxxxx xxxxxx xxxxxxxxxxx; xxx xxxx xx xxxx xx xxxx xxxxx xxxxx xxx xxxxxx",
-  "console.views.device-add.messages.joinEUIPlaceholder": "Xxx xxxxxxxxx Xxxx XXX",
-  "console.views.device-add.messages.leaveBlankPlaceholder": "Xxxxx xxxxx xx xxxxxxxx xxxxxxxxxxxxx",
-  "console.views.device-add.messages.resetsJoinNonces": "Xxxxxx Xxxx Xxxxxx",
-  "console.views.device-add.messages.resetsFCnt": "Xxxxxx Xxxxx Xxxxxxxx",
-  "console.views.device-add.messages.deviceEUIDescription": "Xxx xxxxxx XXX xx xxx xxxxxx xxxxxxxxxx xxx xxxx xxxxxx xx xxx xxxxxxx. Xxx xx xxxxxxx xxxxx.",
-  "console.views.device-add.messages.nwkKeyDescription": "Xxx xxxxxxxxx Xxxxxxx Xxx",
-  "console.views.device-add.messages.appKeyDescription": "Xxx Xxx Xxx xx xxxx xx xxxxxx xxx xxxxxxxxxxxxx xxxxxxx xxxx xxxxxx xxx xxx xxxxxxx.",
-  "console.views.device-add.messages.appSKeyDescription": "Xxx Xxxxxxx Xxx",
-  "console.views.device-add.messages.fwdNtwkKeyDescription": "Xxxxxxxxxx Xxxxxxx Xxxxxxx Xxxxxxxxx Xxx (xx XxXxXXX x.x.x XxxXXxx)",
-  "console.views.device-add.messages.sNtwkSIKeyDescription": "Xxxxxxx Xxxxxxx Xxxxxxx Xxxxxxxxx Xxx (xxxx xxx XxXxXXX x.x+)",
-  "console.views.device-add.messages.ntwkSEncKeyDescription": "Xxxxxxx Xxxxxxx Xxxxxxxxxx Xxx (xxxx xxx XxXxXXX x.x+)",
-  "console.views.device-add.messages.validate8": "Xxxx xxxxx xxxxx xx xx xxxxxxx x xxxxxxxxxx xxxx",
-  "console.views.device-add.messages.validate16": "Xxxx xxxxx xxxxx xx xx xxxxxxx xx xxxxxxxxxx xxxx",
-  "console.views.device-add.messages.validate32": "Xxxx xxxxx xxxxx xx xx xxxxxxx xx xxxxxxxxxx xxxx",
-  "console.views.device-add.messages.supportsClassC": "Xxxxxxxx Xxxxx X",
-  "console.views.device-add.messages.activationMode": "Xxxxxxxxxx Xxxx",
-  "console.views.device-add.messages.otaa": "Xxxx Xxx Xxx Xxxxxxxxxx (XXXX)",
-  "console.views.device-add.messages.abp": "Xxxxxxxxxx Xx Xxxxxxxxxxxxxxx (XXX)",
-  "console.views.device-add.messages.resetWarning": "Xxxxxxxx xx xxxxxxxx xxx xxxxx xxxx xxxxxx xxxxxxxxxxx xxx xxxxxx xxxxxxx",
   "console.views.device-overview.index.hardware": "Xxxxxxxx",
   "console.views.device-overview.index.brand": "Xxxxx",
   "console.views.device-overview.index.model": "Xxxxx",
